# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mobile Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  brand: string;
  category: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  imageUrl: string | null;
  inStock: boolean;
  featured: boolean;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  /** @nullable */
  specs?: string | null;
  /** @nullable */
  badge?: string | null;
}

export interface ProductInput {
  name: string;
  brand: string;
  category: string;
  description?: string;
  price: number;
  originalPrice?: number;
  imageUrl?: string;
  inStock: boolean;
  featured?: boolean;
  rating?: number;
  reviewCount?: number;
  specs?: string;
  badge?: string;
}

export interface ProductUpdate {
  name?: string;
  brand?: string;
  category?: string;
  description?: string;
  price?: number;
  originalPrice?: number;
  imageUrl?: string;
  inStock?: boolean;
  featured?: boolean;
  rating?: number;
  reviewCount?: number;
  specs?: string;
  badge?: string;
}

export type ProductStatsCategoriesItem = {
  name: string;
  count: number;
};

export type ProductStatsBrandsItem = {
  name: string;
  count: number;
};

export interface ProductStats {
  totalProducts: number;
  inStockCount: number;
  categories: ProductStatsCategoriesItem[];
  brands: ProductStatsBrandsItem[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  productCount?: number | null;
}

export interface CartItem {
  id: number;
  productId: number;
  quantity: number;
  product: Product;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface CartSummary {
  itemCount: number;
  subtotal: number;
  discount: number;
  total: number;
}

export type ListProductsParams = {
category?: string;
brand?: string;
search?: string;
featured?: boolean;
};