# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mobile Store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog operations
  - name: cart
    description: Shopping cart operations
  - name: categories
    description: Category operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/featured:
    get:
      operationId: listFeaturedProducts
      tags: [products]
      summary: Get featured products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/stats:
    get:
      operationId: getProductStats
      tags: [products]
      summary: Get product statistics by category
      responses:
        "200":
          description: Product stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductStats"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart items
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItem"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"

  /cart/{id}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  /cart/summary:
    get:
      operationId: getCartSummary
      tags: [cart]
      summary: Get cart totals and item count
      responses:
        "200":
          description: Cart summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartSummary"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Product:
      type: object
      required: [id, name, brand, category, price, imageUrl, inStock, featured]
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        inStock:
          type: boolean
        featured:
          type: boolean
        rating:
          type: ["number", "null"]
        reviewCount:
          type: ["integer", "null"]
        specs:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]

    ProductInput:
      type: object
      required: [name, brand, category, price, inStock]
      properties:
        name:
          type: string
        brand:
          type: string
        category:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        specs:
          type: string
        badge:
          type: string

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        category:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        specs:
          type: string
        badge:
          type: string

    ProductStats:
      type: object
      required: [totalProducts, categories, brands, inStockCount]
      properties:
        totalProducts:
          type: integer
        inStockCount:
          type: integer
        categories:
          type: array
          items:
            type: object
            required: [name, count]
            properties:
              name:
                type: string
              count:
                type: integer
        brands:
          type: array
          items:
            type: object
            required: [name, count]
            properties:
              name:
                type: string
              count:
                type: integer

    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        productCount:
          type: ["integer", "null"]

    CartItem:
      type: object
      required: [id, productId, quantity, product]
      properties:
        id:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        product:
          $ref: "#/components/schemas/Product"

    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer

    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    CartSummary:
      type: object
      required: [itemCount, subtotal, discount, total]
      properties:
        itemCount:
          type: integer
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number