# lib/api-zod/src/generated/types/cartItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mobile Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItemInput {
  productId: number;
  quantity: number;
}