# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mobile Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  brand: string;
  category: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  imageUrl: string | null;
  inStock: boolean;
  featured: boolean;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  /** @nullable */
  specs?: string | null;
  /** @nullable */
  badge?: string | null;
}