# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mobile Store API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  name?: string;
  brand?: string;
  category?: string;
  description?: string;
  price?: number;
  originalPrice?: number;
  imageUrl?: string;
  inStock?: boolean;
  featured?: boolean;
  rating?: number;
  reviewCount?: number;
  specs?: string;
  badge?: string;
}